requirements.txt
aiogram==2.17.1
python-dotenv==0.19.2
splinter==0.16.0